% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createSimulatedStudyPopulation}
\alias{createSimulatedStudyPopulation}
\title{Create a Cyclops-ready tibble for analysis using the `fitOutcomeModel` function.}
\usage{
createSimulatedStudyPopulation(matchedCohort, allOutcomes, outcomeIds)
}
\arguments{
\item{matchedCohort}{A tibble containing the matched cohort table generated by the `generateMatchedCohort` function.}

\item{allOutcomes}{A tibble containing the outcomes for the matched cohort generated by the `getOutcomesForMatchedCohort` function.}

\item{outcomeIds}{A vector of outcome IDs to include in the study population.}
}
\value{
A `tibble` containing the simulated study population data with the following columns:
- `exposureId`: The ID of the exposure cohort.
- `strataId`: The ID of the strata.
- `subjectId`: The ID of the subject.
- `cohortStartDate`: The start date of the target/comparator cohort.
- `timeAtRisk`: The time at risk in days for the subject.
- `y`: The outcome indicator (1 = outcome occurred, 0 = outcome did not occur).
- `daysToEvent`: The number of days to the outcome event from the cohort start date.
}
\description{
Create a Cyclops-ready tibble for analysis using the `fitOutcomeModel` function.
}
