% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outcome.R
\name{createStudyPopulation}
\alias{createStudyPopulation}
\title{Create a study population}
\usage{
createStudyPopulation(concurrentComparatorData, outcomeId)
}
\arguments{
\item{concurrentComparatorData}{An object of type [ConcurrentComparatorData] as created using the
[getDbConcurrentComparatorData] function.}

\item{outcomeId}{The ID of the outcome.}
}
\value{
A `tibble` specifying the study population. This `tibble` will have the following columns:

- `subjectId`: The person sequence ID of the subject.
- `strataId` : The person's stratum
- `exposureId` : 1 == target; 0 == comparator
- `cohortStartdate`: The index date.
- `outcomeCount` The number of outcomes observed during the risk window.
- `timeAtRisk`: The number of days in the risk window.
- `daysToEvent` : The number of days from cohortStartDate on which outcome occurs
}
\description{
Create a study population
}
\details{
TODO
}
