% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{extractMatchedCohortToLocalSystem}
\alias{extractMatchedCohortToLocalSystem}
\title{TODO: doc
Takes andromeda object and extracts matched to local system and returns it}
\usage{
extractMatchedCohortToLocalSystem(andromeda, connection, targetId)
}
\description{
TODO: doc
Takes andromeda object and extracts matched to local system and returns it
}
