% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{extractOutcomesToLocalSystem}
\alias{extractOutcomesToLocalSystem}
\title{TODO: doc}
\usage{
extractOutcomesToLocalSystem(
  andromeda,
  connection,
  outcomeDatabaseSchema,
  cdmDatabaseSchema,
  outcomeTable,
  outcomeIds,
  targetId,
  timeAtRiskStart,
  timeAtRiskEnd
)
}
\description{
TODO: doc
}
