% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{extractStrataToLocalSystem}
\alias{extractStrataToLocalSystem}
\title{TODO: doc
Takes andromeda object and extracts strata to local system and returns it}
\usage{
extractStrataToLocalSystem(connection, targetId, andromeda)
}
\description{
TODO: doc
Takes andromeda object and extracts strata to local system and returns it
}
