% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{filterPersonsAgainstHighRiskGroupDefinition}
\alias{filterPersonsAgainstHighRiskGroupDefinition}
\title{Filters a persons tibble against a high risk group definition.}
\usage{
filterPersonsAgainstHighRiskGroupDefinition(
  persons,
  exposureCohort,
  highRiskGroupDefinition,
  forCriteria = "all"
)
}
\arguments{
\item{persons}{A tibble generated by the `generateSimulatedPersonsData` function.}

\item{exposureCohort}{A cohort against which to apply the risk criteria. This cohort is left-joined with persons.}

\item{highRiskGroupDefinition}{A list of high risk group definitions. Each definition is a list with the following
                                 elements:
                                 - `var`: The variable to filter on. Must be one of `age_group_id`, `gender_concept_id`,
                                   `race_concept_id`, `ethnicity_concept_id`, `cohort_start_date`, or `cohort_end_date`
                                 - `operator`: The operator to use for the filter. Must be one of `=`, `==`, `<`, `<=`, `>`, `>=`,
                                   `in`, or `between`

                                   Note: between requires a 2-item vector and checks the value is between, inclusive of endpoints.
                                 - `val`: The value to filter on. If `operator` is `in`, this should be a vector of
                                    values.}

\item{forCriteria}{A value of `any` or `all` to determine whether risk group must meet all (`all`) criteria or
any one of the criteria (`any`).}
}
\value{
A vector of unique person IDs that match the provided group definition.
}
\description{
Filters a persons tibble against a high risk group definition.
}
