% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outcome.R
\name{fitOutcomeModel}
\alias{fitOutcomeModel}
\title{Create an outcome model, and compute the relative risk}
\usage{
fitOutcomeModel(
  population,
  modelType = "cpr",
  profileGrid = NULL,
  profileBounds = c(log(0.1), log(10)),
  prior = createPrior("laplace", useCrossValidation = TRUE),
  control = createControl(cvType = "auto", seed = 1, resetCoefficients = TRUE,
    startingVariance = 0.01, tolerance = 2e-07, cvRepetitions = 10, noiseLevel = "quiet")
)
}
\arguments{
\item{population}{A population object generated by [createStudyPopulation()],
potentially filtered by other functions.}

\item{modelType}{The type of outcome model that will be used. TODO}

\item{profileGrid}{A one-dimensional grid of points on the log(relative risk) scale where
the likelihood for coefficient of variables is sampled. See details.}

\item{profileBounds}{The bounds (on the log relative risk scale) for the adaptive sampling
of the likelihood function. See details.}

\item{prior}{The prior used to fit the model. See [Cyclops::createPrior()]
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyper-parameters of the prior (if applicable). See
[Cyclops::createControl()] for details.}
}
\value{
An object of class `OutcomeModel`. Generic function `print`, `coef`, and
`confint` are available.
}
\description{
Create an outcome model, and computes the relative risk
}
\details{
For likelihood profiling, either specify the `profileGrid` for a completely user- defined grid, or
`profileBounds` for an adaptive grid. Both should be defined on the log effect size scale. When both
`profileGrid` and `profileGrid` are `NULL` likelihood profiling is disabled.
}
