% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateComparatorCohortFromSimulatedTargetCohortTable}
\alias{generateComparatorCohortFromSimulatedTargetCohortTable}
\title{Generate a comparator cohort from a simulated target cohort table in the same format as
created by the definition encoded in CohortExtraction.sql.}
\usage{
generateComparatorCohortFromSimulatedTargetCohortTable(
  targetCohort,
  riskWindowAfterTargetExposureDays = 21,
  washoutPeriodDays = 22
)
}
\arguments{
\item{targetCohort}{A tibble generated by the `generateTargetCohortFromSimulatedCohortTable`
function.}

\item{riskWindowAfterTargetExposureDays}{The number of days after the target exposure to include in the risk window.}

\item{washoutPeriodDays}{The number of days to shift the cohort start date by.}
}
\value{
A tibble with the following columns:
- `cohort_definition_id`: The ID of the target cohort definition.
- `subject_id`: The person sequence ID of the subject.
- `cohort_start_date`: The date the cohort starts (i.e. the start of the safe window)
- `cohort_end_date`: The date the cohort ends (i.e., the end of the safe window)
}
\description{
Generate a comparator cohort from a simulated target cohort table in the same format as
created by the definition encoded in CohortExtraction.sql.
}
\details{
The Comparator Cohort is generated by taking the latest entry for each person in the target cohort
and shifting the cohort start date by the washout period and the cohort end date by the risk window.
}
