% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateMatchedCohort}
\alias{generateMatchedCohort}
\title{Generates a person-level dataset including both the target and comparator cohort that can be used for analysis.}
\usage{
generateMatchedCohort(targetCohort, comparatorCohort, matchedStrata, persons)
}
\arguments{
\item{targetCohort}{A target cohort data frame generated by `generateTargetCohortFromSimulatedCohortTable` function.}

\item{comparatorCohort}{A comparator cohort data frame generated by `generateComparatorCohortFromSimulatedTargetCohortTable` function.}

\item{matchedStrata}{A matched strata data frame generated by `generateMatchedStrataForComparatorCohort` function.}

\item{persons}{A persons data frame generated by `generateSimulatedPersonsData` function.}
}
\value{
A data frame with the following columns:
- `cohortDefinitionId`: The cohort definition ID for the cohort record.
- `exposureId`: Whether the record is from the target at-risk cohort. (1) or comparator safe cohort (0).
- `strataId`: The strata ID for the cohort record.
- `subjectId`: The subject ID for the cohort record.
- `cohortStartDate`: The cohort start date for the cohort record.
- `cohortEndDate`: The cohort end date for the cohort record.
- `ageGroupId`: The age group ID for the cohort.
- `genderConceptId`: The gender concept ID for the cohort.
- `raceConceptId`: The race concept ID.
- `ethnicityConceptId`: The ethnicity concept ID.
}
\description{
Generates a person-level dataset including both the target and comparator cohort that can be used for analysis.
}
