% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateMatchedStrataForComparatorCohort}
\alias{generateMatchedStrataForComparatorCohort}
\title{Generate matched strata from existing strata and a comparator cohort.}
\usage{
generateMatchedStrataForComparatorCohort(strata, comparatorCohort, persons)
}
\arguments{
\item{strata}{A strata tibble generated by `generateStrataFromCohort` function.}

\item{comparatorCohort}{A comparator cohort tibble generated by
`generateComparatorCohortFromSimulatedTargetCohortTable` function.}

\item{persons}{A persons tibble generated by `generateSimulatedPersonsData` function.}
}
\value{
A tibble with the following columns:
- cohort_definition_id: The ID of the cohort definition.
- strata_id: The ID of the strata, which is unique to each combination of cohort_definition_id,
             cohort_start_date, age_group_id, gender_concept_id, race_concept_id, and ethnicity_concept_id
             in the cohort tibble.
- `cohort_start_date`: The start date of the cohort.
- `age_group_id`: The age group ID as generated by the `getAgeGroupIdForReferenceDateAndDateOfBirth` function
- `gender_concept_id`: The OMOP gender concept ID.
- `race_concept_id`: The OMOP race concept ID.
- `ethnicity_concept_id`: The OMOP ethnicity concept ID.
}
\description{
Generate matched strata from existing strata and a comparator cohort.
}
\details{
This gets all the strata that exist in a comparator cohort and matches them to the strata
in the provided strata tibble. This has effect of remove strata from the provided strata
that do not exist in the comparator cohort.
}
