% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateOutcomesCohortForComparatorCohortFromLastExposureInTargetCohort}
\alias{generateOutcomesCohortForComparatorCohortFromLastExposureInTargetCohort}
\title{Generate outcomes for members of a comparator cohort using the last exposure of a target cohort.}
\usage{
generateOutcomesCohortForComparatorCohortFromLastExposureInTargetCohort(
  targetCohortLastExposuresOnly,
  includePersonIds = c(),
  outcomeId,
  outcomeRate,
  washoutPeriodDays,
  riskWindowAfterTargetExposureDays
)
}
\arguments{
\item{includePersonIds}{A vector of person IDs to include in the outcomes cohort. If empty, all persons in the
comparator cohort will be included.}

\item{outcomeId}{The ID of the outcome cohort definition. TODO: Multiple outcomes?}

\item{outcomeRate}{The proportion of persons in the comparator cohort who will experience the outcome.}

\item{washoutPeriodDays}{The number of days after entering the target cohort until the person is considered no longer
at risk of adverse events caused by exposure to the vaccine.}

\item{riskWindowAfterTargetExposureDays}{The number of days after entering the target cohort when a person is
considered at risk of adverse events caused by exposure to the vaccine.}

\item{comparatorCohort}{A data frame generated by the `getLatestCohortEntryFromEachPersonInTargetCohort` function.}
}
\value{
A data frame containing the outcome cohort with the the following columns:
- `cohort_definition_id`: The ID of the outcome cohort definition.
- `subject_id`: The ID of the subject.
- `cohort_start_date`: The start date of the outcome cohort.
- `cohort_end_date`: The end date of the outcome cohort.
}
\description{
This takes the last exposure of the target cohort and generates outcomes for the comparator cohort. The outcome occurrence date
will be a random day between `washoutPeriodDays` days after the date of the target cohort start date (beginning of safe window)
 and `riskWindowAfterTargetExposureDays` days after the beginning of the safe window.
}
