% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateSimulatedConceptAncestorTable}
\alias{generateSimulatedConceptAncestorTable}
\title{Generate a simulated concept ancestor table from a condition era table.}
\usage{
generateSimulatedConceptAncestorTable(conditionEraTable)
}
\arguments{
\item{conditionEraTable}{A tibble generated by the `generateSimulatedConditionEraTable` function.}
}
\value{
A tibble.
}
\description{
Generate a simulated concept ancestor table from a condition era table.
}
