% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateSimulatedConditionEraTable}
\alias{generateSimulatedConditionEraTable}
\title{Generate a simulated condition era table from a cohort table and set of outcome IDs.}
\usage{
generateSimulatedConditionEraTable(cohort, outcomeIds)
}
\arguments{
\item{cohort}{A tibble generated by the `generateSimulatedCohortTable` function.}

\item{outcomeIds}{A vector of IDs of the adverse event outcomes that are being studied.}
}
\value{
A tibble.
}
\description{
Generate a simulated condition era table from a cohort table and set of outcome IDs.
}
