% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateSimulatedObservationPeriodTable}
\alias{generateSimulatedObservationPeriodTable}
\title{Generate a simulated observation period table consistent with the OMOP CDM.}
\usage{
generateSimulatedObservationPeriodTable(cohort)
}
\arguments{
\item{cohort}{A tibble generated by `generateSimulatedCohortTable`.}
}
\value{
A tibble with the following columns:
- `observation_period_id`: A unique identifier for each observation period.
- `person_id`: A unique identifier for each person.
- `observation_period_start_date`: The start date of the observation period.
- `observation_period_end_date`: The end date of the observation period.
}
\description{
Generate a simulated observation period table consistent with the OMOP CDM.
}
