% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateSimulatedPersonsData}
\alias{generateSimulatedPersonsData}
\title{Generate a simulated persons table consistent with the OMOP CDM.}
\usage{
generateSimulatedPersonsData(n = 1000, studyEndDate = "2021-06-01")
}
\arguments{
\item{n}{The number of persons to generate.}

\item{studyEndDate}{The date the study ended in YYYY-MM-DD format.}
}
\value{
A tibble with the following columns:

- `person_id`: A unique identifier for each person.
- `birth_datetime`: The date of birth of the person in YYYY-MM-DD format.
- `gender_concept_id`: The OMOP CDM gender concept ID.
- `race_concept_id`: The OMOP CDM race concept ID.
- `ethnicity_concept_id`: The OMOP CDM ethnicity concept ID.
}
\description{
Generate a simulated persons table consistent with the OMOP CDM.
}
