% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateSimulatedPopulationAnalysisData}
\alias{generateSimulatedPopulationAnalysisData}
\title{Generate simulated population data for target and comparator suitable for analysis.}
\usage{
generateSimulatedPopulationAnalysisData(
  targetId,
  outcomeIds,
  n = 1e+05,
  proportionSecondShot = 0.65,
  timeAtRiskStartDays = 1,
  timeAtRiskEndDays = 21,
  washoutPeriodDays = 22,
  highRiskGroupDefinition = list(),
  matchHighRiskCriteria = "any",
  lowRiskComparatorIncidence = 1e-04,
  lowRiskTargetIncidence = 1e-04,
  highRiskComparatorIncidence = 1e-04,
  highRiskTargetIncidence = 1e-04,
  studyStartDate = "2020-12-18",
  studyEndDate = "2021-06-30"
)
}
\arguments{
\item{targetId}{The ID of the target cohort.}

\item{outcomeIds}{The IDs of the outcomes.}

\item{n}{The number of subjects in the simulated study population.}

\item{proportionSecondShot}{The proportion of subjects to receive a  second shot.}

\item{timeAtRiskEndDays}{The number of days after target exposure to consider subject at risk.}

\item{washoutPeriodDays}{The number of days after entering target cohort to consider subject
'"safe" and eligible for comparator cohort.}

\item{highRiskGroupDefinition}{A list to pass to the `filterPersonsAgainstHighRiskGroupDefinition()`
defining which subjects are considered high risk.}

\item{matchHighRiskCriteria}{A value of `any` (has any high risk characteristic) or `all` (has all
defined high risk characteristics) to
define how to match high risk subjects.}

\item{lowRiskComparatorIncidence}{The incidence rate of the outcome in the low risk comparator group
(those who do not match the high risk criteria).}

\item{lowRiskTargetIncidence}{The incidence rate of the outcome in the low risk target group (those
who do not match the high risk criteria).}

\item{highRiskComparatorIncidence}{The incidence rate of the outcome in the high risk comparator group
(those who match the high risk criteria).}

\item{highRiskTargetIncidence}{The incidence rate of the outcome in the high risk target group (those
who match the high risk criteria).}

\item{studyStartDate}{The start date of the study in the format 'YYYY-MM-DD'.}

\item{studyEndDate}{The end date of the study in the format 'YYYY-MM-DD'.}
}
\value{
A `tibble` containing the simulated study population data with the following columns:
exposureId strataId subjectId cohortStartDate timeAtRisk     y daysToEvent
- `exposureId`: 1 == target; 0 == comparator
- `strataId`: The stratum of the entry.
- `subjectId`: The subjectId of the person from the simulated `persons` table.
- `cohortStartDate`: The date the subject entered the cohort in the format 'YYYY-MM-DD'.
- `timeAtRisk`: The number of days in the risk window.
- `y`: The outcome status (1 = outcome occurred; 0 = outcome did not occur).
- `daysToEvent`: The number of days from cohortStartDate on which outcome occurs; NA if no outcome occured.
}
\description{
Generate simulated population data for target and comparator suitable for analysis.
}
