% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateStrataFromCohort}
\alias{generateStrataFromCohort}
\title{Gets unique strata from a cohort and persond tibble.}
\usage{
generateStrataFromCohort(cohort, persons)
}
\arguments{
\item{cohort}{A cohort tibble with cohort_definition_id, cohort_start_date}

\item{persons}{A persons tibble generated by `generateSimulatedPersonsData` function.}
}
\value{
A tibble with the following columns:
- cohort_definition_id: The ID of the cohort definition.
- strata_id: The ID of the strata, which is unique to each combination of cohort_definition_id,
             cohort_start_date, age_group_id, gender_concept_id, race_concept_id, and ethnicity_concept_id
             in the cohort tibble.
- cohort_start_date: The start date of the cohort.
- age_group_id: The age group ID as generated by the `getAgeGroupIdForReferenceDateAndDateOfBirth` function
- gender_concept_id: The OMOP gender concept ID.
- race_concept_id: The OMOP race concept ID.
- ethnicity_concept_id: The OMOP ethnicity concept ID.
}
\description{
Gets unique strata from a cohort and persond tibble.
}
