% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateTargetCohortFromPersons}
\alias{generateTargetCohortFromPersons}
\title{Generate a target cohort from a persons tibble.}
\usage{
generateTargetCohortFromPersons(
  persons,
  targetId,
  washoutPeriodDays,
  studyEndDate,
  targetDaysEligibleBeforeStudyEndDate,
  proportionSecondShot,
  daysBetweenShots = 21
)
}
\arguments{
\item{persons}{A tibble generated by the `generateSimulatedPersonsData` function.}

\item{targetId}{The ID of the target cohort definition.}

\item{washoutPeriodDays}{The number of days after entering the target cohort until the person is
considered no longer
at risk of adverse events caused by exposure to the vaccine.}

\item{studyEndDate}{The date the study ended in YYYY-MM-DD format.}

\item{targetDaysEligibleBeforeStudyEndDate}{The number of days before the study end date that a person is eligible
to enter the target cohort.}

\item{proportionSecondShot}{The proportion of people in the target cohort who receive a second shot.}
}
\value{
A tibble with the following columns:
- `cohort_definition_id`
- `subject_id`
- `cohort_start_date`
- `cohort_end_date`
}
\description{
Generate a target cohort from a persons tibble.
}
