% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateTargetCohortFromSimulatedCohortTable}
\alias{generateTargetCohortFromSimulatedCohortTable}
\title{Generate a target cohort from a simulated cohort table in the same format as
created by the definition encoded in CohortExtraction.sql.}
\usage{
generateTargetCohortFromSimulatedCohortTable(
  cohort,
  targetCohortId,
  timeAtRiskEndDays = 21
)
}
\arguments{
\item{cohort}{A tibble generated by the `generateSimulatedCohortTable` function.}

\item{targetCohortId}{The ID of the target cohort definition.}

\item{timeAtRiskEndDays}{The number of days after the target exposure to include in the risk window.}
}
\value{
A tibble with the following columns:
- `cohort_definition_id`: The ID of the target cohort definition.
- `subject_id`: The person sequence ID of the subject.
- `cohort_start_date`: The date the cohort starts (i.e. the start of the risk window)
- `cohort_end_date`: The date the cohort ends (i.e., the end of the risk window)
}
\description{
Generate a target cohort from a simulated cohort table in the same format as
created by the definition encoded in CohortExtraction.sql.
}
