% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getAgeGroupCategories}
\alias{getAgeGroupCategories}
\title{Get random birth dates from age categories.}
\usage{
getAgeGroupCategories()
}
\arguments{
\item{ageCategories}{A vector of age category IDs.}

\item{studyEndDate}{The date the study ended in YYYY-MM-DD format.}
}
\value{
A vector of dates of birth in YYYY-MM-DD format consistent with the provided
study start date.
}
\description{
Get random birth dates from age categories.
}
