% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getAgeGroupIdForReferenceDateAndDateOfBirth}
\alias{getAgeGroupIdForReferenceDateAndDateOfBirth}
\title{Gets an age group ID from a reference date and date of birth.}
\usage{
getAgeGroupIdForReferenceDateAndDateOfBirth(referenceDate, dateOfBirth)
}
\arguments{
\item{referenceDate}{The date to use as the reference date for calculating age.}

\item{dateOfBirth}{The date of birth of the person.}
}
\value{
A numeric value corresponding to the bin number a person would fall in based on their age, with bins of 5 years
starting at birth.
}
\description{
Gets an age group ID from a reference date and date of birth.
}
