% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{getDbConcurrentComparatorData}
\alias{getDbConcurrentComparatorData}
\title{Get the concurrent comparator from the server}
\usage{
getDbConcurrentComparatorData(
  connectionDetails,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  targetId = 1,
  outcomeIds,
  overwriteComparators = FALSE,
  studyStartDate = "",
  studyEndDate = "",
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable,
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable,
  timeAtRiskStart,
  timeAtRiskEnd,
  washoutTime = timeAtRiskEnd + 1,
  intermediateFileNameStem = NULL
)
}
\arguments{
\item{connectionDetails}{An R object of type `connectionDetails` created using the
[DatabaseConnector::createConnectionDetails()] function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
.                                    emulate temp tables, provide a schema with write privileges where temp tables
                                   can be created.}

\item{targetId}{A unique identifier to define the target cohort. targetId is
used to select the COHORT_DEFINITION_ID in the cohort-like table.}

\item{outcomeIds}{A list of cohort IDs used to define outcomes.}

\item{overwriteComparators}{Allow regeneration if comparators if comparatorId already exists
TODO what is the table name?}

\item{studyStartDate}{A calendar date specifying the minimum date that a cohort index
date can appear. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date that a cohort index
date can appear. Date format is 'yyyymmdd'. Important: the study
end data is also used to truncate risk windows, meaning no
outcomes beyond the study end date will be considered.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the
exposure data used to define the exposure cohorts is available.}

\item{exposureTable}{The tablename that contains the exposure cohorts has the
format of a COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the
data used to define the outcome cohorts is available.}

\item{outcomeTable}{The tablename that contains the outcome cohorts has the format of a COHORT table: COHORT_DEFINITION_ID,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{timeAtRiskStart}{The time-at-risk start in days after subject index date.}

\item{timeAtRiskEnd}{The time-at-risk end in days after subject index date.}

\item{washoutTime}{Washout time in days between target and comparator periods}

\item{comparatorId}{A unique identifier to define the comparator cohort. comparatorId
is used to select the COHORT_DEFINITION_ID in the cohort-like
table. TODO UPDATE}
}
\value{
A [ConcurrentComparatorData] object.
}
\description{
This function executes a large set of SQL statements against the database in OMOP CDM format to
extract the data needed to perform the analysis.
}
\details{
TODO
}
