% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getLatestCohortEntryFromEachPersonInTargetCohort}
\alias{getLatestCohortEntryFromEachPersonInTargetCohort}
\title{Gets the latest cohort entry from each person in the target cohort.}
\usage{
getLatestCohortEntryFromEachPersonInTargetCohort(targetCohort)
}
\arguments{
\item{targetCohort}{A data frame generated by the `generateTargetCohortFromPersons` function.}
}
\value{
A data frame with the following columns:
- `cohort_definition_id`
- `subject_id`
- `cohort_start_date`
- `cohort_end_date`
}
\description{
Gets the latest cohort entry from each person in the target cohort.
}
