% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getOutcomesForMatchedCohort}
\alias{getOutcomesForMatchedCohort}
\title{Get the adverse event outcomes for a matched cohort from the cohort table.}
\usage{
getOutcomesForMatchedCohort(
  cohort,
  matchedCohort,
  targetId,
  outcomeIds,
  daysFromObsStart = 1,
  daysToObsEnd = 21,
  studyEndDate = ""
)
}
\arguments{
\item{cohort}{A data frame containing the cohort table generated by the `generateSimulatedCohortTable` function.}

\item{matchedCohort}{A data frame containing the matched cohort table generated by the `generateMatchedCohort` function.}

\item{targetId}{The target cohort definition ID.}

\item{outcomeIds}{A vector of outcome cohort definition IDs.}

\item{daysFromObsStart}{The number of days after the target cohort start date to begin the risk window for adverse events.}

\item{daysToObsEnd}{The number of days after the target cohort start date to end the risk window for adverse events.}

\item{studyEndDate}{The date to use as the end of the study period. If specified, outcomes that occur after this date are excluded.}
}
\value{
A tibble with the following columns:
- `subjectId`: The subject ID.
- `strataId`: The strata ID.
- `outcomeId`: The outcome cohort definition ID.
- `exposureId`: The exposure (i.e., target) cohort definition ID.
- `cohortStartDate`: The start date of the cohort.
- `daysToEvent`: The number of days from the cohort start date to the outcome event.
- `y`: A binary indicator of whether the outcome event occurred within the risk window.
}
\description{
This function takes matched cohort data frame and finds all outcomes matching the specified `outcomeIds` that
occur within the risk window for each record. The risk window is defined using the `daysFromObsStart` and
`daysToObsEnd` parameters. If specified, outcomes are truncated by excluding records that occur after
`studyEndDate.`
}
