% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getPersonLevelDataFromCohortAndStrata}
\alias{getPersonLevelDataFromCohortAndStrata}
\title{Gets person-level dataset that contains only those cohort records that belong to the matched strata.}
\usage{
getPersonLevelDataFromCohortAndStrata(cohort, cohortType, strata, persons)
}
\arguments{
\item{cohort}{A target cohort data frame generated by `generateTargetCohortFromSimulatedCohortTable` function or
a comparator cohort data frame generated by `generateComparatorCohortFromSimulatedTargetCohortTable` function.}

\item{persons}{}
}
\description{
This function takes a cohort, exposureId, strata data frame, and persons data frame and returns a person-level
dataset that can be used for analysis.
}
