% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getRandomAgeCategories}
\alias{getRandomAgeCategories}
\title{Get test vocabulary concepts using OMOP vocabulary.}
\usage{
getRandomAgeCategories(n = 1)
}
\value{
A tibble with columns concept_id, name, domain, and probability.

- concept_id: The unique identifier for the concept.
- name: The name of the concept.
- domain: The domain of the concept.
- prob: The probability of the concept, conditional on the domain.

Probabilities roughly match the 2020 US Census data, with some exceptions.
Race, for example, will not sum to 1.0 because the Census allows for
respondents to select more than one response category.
}
\description{
Get test vocabulary concepts using OMOP vocabulary.
}
