% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getRandomDatesOfBirthFromAgeCategories}
\alias{getRandomDatesOfBirthFromAgeCategories}
\title{Get random dates of birth from a set of age categories.}
\usage{
getRandomDatesOfBirthFromAgeCategories(
  ageCategories,
  studyEndDate = "2021-06-01"
)
}
\arguments{
\item{n}{The number of categories to return.}
}
\value{
A vector of random birthdays in YYYY-MM-DD format.
}
\description{
Get random dates of birth from a set of age categories.
}
