% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getRandomEthnicityCategories}
\alias{getRandomEthnicityCategories}
\title{Get random age categories in five-year bins, roughly representative of the US
population in 2020.}
\usage{
getRandomEthnicityCategories(n = 1)
}
\arguments{
\item{n}{The number of categories to return.}
}
\value{
A vector of age category IDs.
}
\description{
Get random age categories in five-year bins, roughly representative of the US
population in 2020.
}
