% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getRandomGenderCategories}
\alias{getRandomGenderCategories}
\title{Get age group categories used to stratify the analysis population.}
\usage{
getRandomGenderCategories(n = 1)
}
\value{
A data frame with the following columns:

- age_group_id: The unique identifier for the age group.
- name: The name of the age group.
- prob: The probability of the age group.

Probabilities roughly match the 2020 US Census data age reported in Table 2
in the following report, with some rounding and adjustment for bins that
do not align with five-year age groups used in this package.

https://www2.census.gov/library/publications/decennial/2020/census-briefs/c2020br-06.pdf
}
\description{
Get age group categories used to stratify the analysis population.
}
