% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getRandomRaceCategories}
\alias{getRandomRaceCategories}
\title{Get random race categories representative of the US population in 2020.}
\usage{
getRandomRaceCategories(n = 1)
}
\arguments{
\item{n}{The number of categories to return.}
}
\value{
A vector of OMOP CDM race category concept IDs.
}
\description{
Get random race categories representative of the US population in 2020.
}
