% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getTestVocabConcepts}
\alias{getTestVocabConcepts}
\title{Get random ethnicity categories representative of the US population in 2020.}
\usage{
getTestVocabConcepts()
}
\arguments{
\item{n}{The number of categories to return.}
}
\value{
A vector of OMOP CDM ethnicity category concept IDs.
}
\description{
Get random ethnicity categories representative of the US population in 2020.
}
