% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{initializeConcurrentComparatorData}
\alias{initializeConcurrentComparatorData}
\title{Initialize concurrent comparator data as an empty Andromeda object.}
\usage{
initializeConcurrentComparatorData()
}
\value{
A [ConcurrentComparatorData] object.
}
\description{
Initialize concurrent comparator data as an empty Andromeda object.
}
