% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{initializeDatabaseConnection}
\alias{initializeDatabaseConnection}
\title{Initialize database connection}
\usage{
initializeDatabaseConnection(connectionDetails)
}
\arguments{
\item{connectionDetails}{An R object of type `connectionDetails` created using the
[DatabaseConnector::createConnectionDetails()] function.}
}
\value{
A database connection object.
}
\description{
Initialize database connection
}
