% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{loadConcurrentComparatorData}
\alias{loadConcurrentComparatorData}
\title{Load the concurrent comparator data from a file}
\usage{
loadConcurrentComparatorData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class [ConcurrentComparatorData].
}
\description{
Loads an object of type [ConcurrentComparatorData] from a file in the file system.
}
