% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{removeMatchedCohortEpisodesWithZeroTimeAtRisk}
\alias{removeMatchedCohortEpisodesWithZeroTimeAtRisk}
\title{Remove matched cohort episodes with zero time-at-risk.}
\usage{
removeMatchedCohortEpisodesWithZeroTimeAtRisk(concurrentComparatorData)
}
\arguments{
\item{concurrentComparatorData}{A [ConcurrentComparatorData] object.}
}
\value{
A [ConcurrentComparatorData] object.
}
\description{
Remove matched cohort episodes with zero time-at-risk.
}
