% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{saveConcurrentComparatorData}
\alias{saveConcurrentComparatorData}
\title{Save the concurrent comparator data to file}
\usage{
saveConcurrentComparatorData(concurrentComparatorData, file)
}
\arguments{
\item{concurrentComparatorData}{An object of type [ConcurrentComparatorData] as created using the
[getDbConcurrentComparatorData] function.}

\item{file}{The name of the file where the data will be written. If the file already
exists it will be overwritten.}
}
\value{
Returns no output.
}
\description{
Saves an object of type [ConcurrentComparatorData] to a file.
}
