% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{saveIntermediateFile}
\alias{saveIntermediateFile}
\title{Save intermediate file.}
\usage{
saveIntermediateFile(
  concurrentComparatorData,
  intermediateFileNameStem,
  suffix = 1
)
}
\arguments{
\item{concurrentComparatorData}{A [ConcurrentComparatorData] object.}

\item{intermediateFileNameStem}{The stem of the file name to save the intermediate file to.}

\item{suffix}{A suffix to append to the file name.}
}
\value{
A [ConcurrentComparatorData] object.
}
\description{
Save intermediate file.
}
