% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{summarizeSimulatedTestData}
\alias{summarizeSimulatedTestData}
\title{Prints a tabulation of each covariate in the simulated persons data.}
\usage{
summarizeSimulatedTestData(persons, studyEndDate = "2021-06-01")
}
\arguments{
\item{persons}{A tibble generated by the `generateSimulatedPersonsData` function.}

\item{studyEndDate}{The date the study ended in YYYY-MM-DD format.}
}
\value{
Void.
}
\description{
Prints a tabulation of each covariate in the simulated persons data.
}
