% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{translateCohortExtractionSql}
\alias{translateCohortExtractionSql}
\title{TODO: document}
\usage{
translateCohortExtractionSql(
  dbms,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  timeAtRiskEnd,
  washoutTime,
  targetId,
  testing = testing
)
}
\description{
TODO: document
}
