% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeMetaDataToConcurrentComparatorData}
\alias{writeMetaDataToConcurrentComparatorData}
\title{Title}
\usage{
writeMetaDataToConcurrentComparatorData(
  concurrentComparatorData,
  targetId,
  outcomeIds,
  matchedCohortDiagnostics
)
}
\arguments{
\item{concurrentComparatorData}{An Andromeda object returned by `initializeConcurrentComparatorData()`}

\item{targetId}{The ID of the target cohort.}

\item{outcomeIds}{A vector of IDs of the outcomes.}

\item{matchedCohortDiagnostics}{A list of diagnostics from the matched cohort, including acounts of `zeroT` and
`zeroC`, which are the number of subjects with zero time-at-risk in the target
and comparator cohorts, respectively.}
}
\value{
An object of class `ConcurrentComparatorData` with metadata attached.
}
\description{
Title
}
