% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeOutcomesToConcurrentComparatorData}
\alias{writeOutcomesToConcurrentComparatorData}
\title{TODO: doc}
\usage{
writeOutcomesToConcurrentComparatorData(
  concurrentComparatorData,
  connection,
  dbms,
  outcomeDatabaseSchema,
  cdmDatabaseSchema,
  outcomeTable,
  outcomeIds,
  targetId,
  timeAtRiskStart,
  timeAtRiskEnd
)
}
\description{
TODO: doc
}
