% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeStrataToConncurentComparatorData}
\alias{writeStrataToConncurentComparatorData}
\title{TODO: doc
Takes andromeda object and extracts strata to local system and returns it}
\usage{
writeStrataToConncurentComparatorData(
  concurrentComparatorData,
  connection,
  targetId
)
}
\description{
TODO: doc
Takes andromeda object and extracts strata to local system and returns it
}
