% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outcome.R
\name{createStudyPopulation}
\alias{createStudyPopulation}
\title{Create a study population}
\usage{
createStudyPopulation(ConcurrentComparatorData, outcomeId)
}
\arguments{
\item{outcomeId}{The ID of the outcome.}

\item{concurrentComparatorData}{An object of type [ConcurrentComparatorData] as created using the
[getDbConcurrentComparatorData] function.}
}
\value{
A `tibble` specifying the study population. This `tibble` will have the following columns:

- `rowId`: A unique identifier for an exposure.
- `personSeqId`: The person sequence ID of the subject.
- `cohortStartdate`: The index date.
- `outcomeCount` The number of outcomes observed during the risk window.
- `timeAtRisk`: The number of days in the risk window.
- `survivalTime`: The number of days until either the outcome or the end of the risk window.
}
\description{
Create a study population
}
\details{
TODO
}
