% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generateOutcomesCohortFromTargetCohort}
\alias{generateOutcomesCohortFromTargetCohort}
\title{Generate outcomes for members of a target cohort.}
\usage{
generateOutcomesCohortFromTargetCohort(
  targetCohort,
  includePersonIds = c(),
  outcomeId,
  outcomeRate,
  riskWindowAfterTargetExposureDays
)
}
\arguments{
\item{targetCohort}{A data frame generated by the `getLatestCohortEntryFromEachPersonInTargetCohort` function.}

\item{includePersonIds}{A vector of person IDs to include in the outcomes cohort. If empty, all persons in the
target cohort will be included.}

\item{outcomeId}{The ID of the outcome cohort definition. TODO: Multiple outcomes?}

\item{outcomeRate}{The proportion of persons in the comparator cohort who will experience the outcome.}

\item{riskWindowAfterTargetExposureDays}{The number of days after the target exposure that the adverse event outcome can occur.}
}
\value{
A data frame containing the outcome cohort with the the following columns:
- `cohort_definition_id`
- `subject_id`
- `cohort_start_date`
- `cohort_end_date`
}
\description{
Generate outcomes for members of a target cohort.
}
