% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{getPersonLevelDataFromCohortAndStrata}
\alias{getPersonLevelDataFromCohortAndStrata}
\title{Gets person-level dataset that contains only those cohort records that belong to the matched strata.}
\usage{
getPersonLevelDataFromCohortAndStrata(cohort, cohortType, strata, persons)
}
\arguments{
\item{cohort}{A target cohort tibble generated by `generateTargetCohortFromSimulatedCohortTable` function or
a comparator cohort tibble generated by `generateComparatorCohortFromSimulatedTargetCohortTable` function.}

\item{strata}{A matched strata tibble generated by `generateMatchedStrataForComparatorCohort` function.}

\item{persons}{A persons tibble generated by `generateSimulatedPersonsData` function.}

\item{exposureId}{The exposure ID for the cohort.}
}
\value{
A tibble with the following columns:
- `cohort_definition_id`: The cohort definition ID.
- `exposure_id`: Whether the record is from the target cohort (1) or comparator cohort (0).
- `strata_id`: The strata ID for the cohort record.
- `subject_id`: The subject ID for the cohort record.
- `cohort_start_date`: The cohort start date.
- `cohort_end_date`: The cohort end date.
- `age_group_id`: The age group ID for the record
- `gender_concept_id`: The gender concept ID for the record.
- `race_concept_id`: The race concept ID for the record.
- `ethnicity_concept_id`: The ethnicity concept ID for the record.
}
\description{
This function takes a cohort, exposureId, strata tibble, and persons tibble and returns a person-level
dataset that can be used for analysis.
}
