% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{removeMatchedCohortEpisodesWithZeroTimeAtRisk}
\alias{removeMatchedCohortEpisodesWithZeroTimeAtRisk}
\title{TODO: doc}
\usage{
removeMatchedCohortEpisodesWithZeroTimeAtRisk(concurrentComparatorData)
}
\description{
TODO: doc
}
