% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{translateCohortExtractionSql}
\alias{translateCohortExtractionSql}
\title{Translate CohortExtraction.sql}
\usage{
translateCohortExtractionSql(
  dbms,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  timeAtRiskEnd,
  washoutTime,
  targetId,
  testing = FALSE
)
}
\arguments{
\item{dbms}{The database management system understood by SqlRender.}

\item{cdmDatabaseSchema}{The name of the database schema where the OMOP CDM instance is stored.}

\item{cohortDatabaseSchema}{The name of the database schema where the cohort data is stored.}

\item{cohortTable}{The name of the table where the cohort data is stored.}

\item{timeAtRiskEnd}{The time-at-risk end in days after subject index date.}

\item{washoutTime}{Washout time in days between target and comparator periods.}

\item{targetId}{The ID of the target cohort.}

\item{testing}{Whether to run in testing mode.}
}
\value{
A string of SQL.
}
\description{
Translate CohortExtraction.sql
}
