% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeMatchedCohortsToScratchDatabase}
\alias{writeMatchedCohortsToScratchDatabase}
\title{TODO: document once working}
\usage{
writeMatchedCohortsToScratchDatabase(
  connectionDetails,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  timeAtRiskEnd,
  washoutTime,
  targetId
)
}
\description{
TODO: document once working
}
