% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeMatchedCohortToConcurrentComparatorData}
\alias{writeMatchedCohortToConcurrentComparatorData}
\title{TODO: doc
Takes andromeda object and extracts matched to local system and returns it}
\usage{
writeMatchedCohortToConcurrentComparatorData(
  concurrentComparatorData,
  connection,
  dbms,
  targetId
)
}
\description{
TODO: doc
Takes andromeda object and extracts matched to local system and returns it
}
