% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeMetaDataToConcurrentComparatorData}
\alias{writeMetaDataToConcurrentComparatorData}
\title{TODO: doc}
\usage{
writeMetaDataToConcurrentComparatorData(
  concurrentComparatorData,
  targetId,
  outcomeIds,
  matchedCohortDiagnostics
)
}
\description{
TODO: doc
}
