% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeOutcomesToConcurrentComparatorData}
\alias{writeOutcomesToConcurrentComparatorData}
\title{Write outcomes to concurrent comparator data.}
\usage{
writeOutcomesToConcurrentComparatorData(
  concurrentComparatorData,
  connection,
  dbms,
  outcomeDatabaseSchema,
  cdmDatabaseSchema,
  outcomeTable,
  outcomeIds,
  targetId,
  timeAtRiskStart,
  timeAtRiskEnd,
  testing = FALSE
)
}
\arguments{
\item{concurrentComparatorData}{A [ConcurrentComparatorData] object.}

\item{connection}{A database connection object.}

\item{dbms}{The database management system of the connection understood by SqlRender.}

\item{outcomeDatabaseSchema}{The name of the database schema where outcomes are stored.}

\item{cdmDatabaseSchema}{The name of the database schema where the OMOP CDM instance is stored.}

\item{outcomeTable}{The name of the table where outcomes are stored.}

\item{outcomeIds}{A vector of IDs of the outcomes.}

\item{targetId}{The ID of the target cohort.}

\item{timeAtRiskStart}{The time-at-risk start in days after subject index date.}

\item{timeAtRiskEnd}{The time-at-risk end in days after subject index date.}

\item{testing}{Whether to run in testing mode.}
}
\value{
A [ConcurrentComparatorData] object.
}
\description{
Write outcomes to concurrent comparator data.
}
