% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{writeStrataToConncurentComparatorData}
\alias{writeStrataToConncurentComparatorData}
\title{Write strata to concurrent comparator data.}
\usage{
writeStrataToConncurentComparatorData(
  concurrentComparatorData,
  connection,
  dbms,
  targetId
)
}
\arguments{
\item{concurrentComparatorData}{A [ConcurrentComparatorData] object.}

\item{connection}{A database connection object.}

\item{dbms}{The database management system of the connection understood by SqlRender.}

\item{targetId}{The ID of the target cohort.}
}
\value{
A [ConcurrentComparatorData] object.
}
\description{
Write strata to concurrent comparator data.
}
